+++ src/st_stuff.c
@@ -93,6 +93,7 @@ static boolean use_nughud;
 static boolean hud_blink_keys;
 static boolean sts_show_berserk;
 int force_carousel;
+int carousel_fadeout;
 
 // Font extras ---------------------------------------------------------------
 
@@ -4011,10 +4012,14 @@ void ST_BindSTSVariables(void)
   M_BindBool("weapon_carousel", &weapon_carousel, NULL,
              true, ss_weap, wad_no, "Show weapon carousel");
 
-  // [Nugget]
+  // [Nugget] ----------------------------------------------------------------
+
   M_BindNum("force_carousel", &force_carousel, NULL,
             1, 0, 2, ss_weap, wad_no,
             "Force display of weapon carousel (0 = Off; 1 = Off player; 2 = Always)");
+
+  M_BindBool("carousel_fadeout", &carousel_fadeout, NULL,
+             true, ss_none, wad_yes, "Carousel fades out");
 }
 
 //----------------------------------------------------------------------------